+++ scripts/spells/divine-smite.js
@@ -5,7 +5,7 @@
 
     Using a Higher-Level Spell Slot. The damage increases by 1d8 for each spell slot level above 1.
 */
-const version = "12.4.1";
+const version = "14.5.0";
 const optionName = "Divine Smite";
 
 try {
@@ -13,7 +13,7 @@ try {
         let targetToken = workflow.hitTargets.first();
         let undead = ["undead", "fiend"].some(type => (targetToken.actor.system.details.type?.value || "").toLowerCase().includes(type));
         const spellLevel = actor.flags["fvtt-trazzm-homebrew-5e"].DivineSmite.level ?? 1;
-        const diceCount = 1 + spellLevel + (undead ? 1 : 0);
+        const diceCount = 1 + Number(spellLevel) + (undead ? 1 : 0);
 
         return new game.system.dice.DamageRoll(`${diceCount}d8`, {}, {
             isCritical: workflow.isCritical,