+++ scripts/spells/protection-from-poison.js
@@ -3,13 +3,22 @@
 	throws to avoid or end the Poisoned condition, and it has Resistance to Poison damage.
 */
 const optionName = "Protection from Poison";
-const version = "12.4.0";
+const version = "14.5.0";
 
 try {
     const targetToken = workflow.targets.first();
+    if (args[0].macroPass === "postActiveEffects" && targetToken) {
+        if (actor.statuses.has('poisoned')) {
+            // remove the effects that contribute the removed condition
+            let actorEffects = Array.from(targetToken.actor.allApplicableEffects());
+            let removedEffects = actorEffects.filter(s => s.statuses.has('poisoned'));
+            if (removedEffects) {
+                const effectIdList = removedEffects.map(obj => obj.id);
+                await MidiQOL.socket().executeAsGM("removeEffects", {actorUuid: targetToken.actor.uuid, effects: effectIdList});
+            }
 
-    if (args[0].macroPass === "postActiveEffects") {
-        await targetToken.actor.toggleStatusEffect('poisoned', {active: false});
+            await targetToken.actor.toggleStatusEffect('poisoned', {active: false});
+        }
     }
 
 } catch (err) {